test_project/requirements.txt
decorator==4.0.10
Django==1.10.5
django-debug-toolbar==1.6
django-extensions==1.7.5
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2016.10
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.15
